role and permission set

+++ app/Http/Controllers/Admin/AnnouncementController.php
@@ -11,6 +11,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class AnnouncementController extends Controller
 {
+	public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-announcement')->only('index');
+        $this->middleware('permission:create-announcement')->only(['create', 'store']);
+        $this->middleware('permission:edit-announcement')->only(['edit', 'update']);
+        $this->middleware('permission:delete-announcement')->only('destroy');
+    }
 	/**
      * Display a listing of the resource.
      */
@@ -30,18 +44,31 @@ class AnnouncementController extends Controller
 					$editUrl = route('announcements.edit', $announcement->id);
 					$deleteUrl = route('announcements.destroy', $announcement->id);
 
-					return '
-						<a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
-							<i class="fa fa-edit"></i>
-						</a>
+					$buttons = '';
 
-						<form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $announcement->id . '" data-name="' . e($announcement->title) . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </form>
-					';
+					if (auth()->user()->can('edit-announcement')) {
+						$buttons .= '
+							<a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
+								<i class="fa fa-edit"></i>
+							</a>
+						';
+					}
+
+					if (auth()->user()->can('delete-announcement')) {
+						$buttons .= '
+							<form action="' . $deleteUrl . '" method="POST" 
+								class="delete-form d-inline" 
+								data-id="' . $announcement->id . '" 
+								data-name="' . e($announcement->title) . '">
+								' . csrf_field() . method_field('DELETE') . '
+								<button type="submit" class="btn btn-danger btn-sm deleteItem">
+									<i class="fa fa-trash"></i>
+								</button>
+							</form>
+						';
+					}
+
+					return $buttons;
 				})
                 ->rawColumns(['status', 'action'])
                 ->make(true);
