+++ app/Http/Controllers/Admin/PageController.php
@@ -13,6 +13,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class PageController extends Controller
 {
+    public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-page')->only('index');
+        $this->middleware('permission:create-page')->only(['create', 'store']);
+        $this->middleware('permission:edit-page')->only(['edit', 'update']);
+        $this->middleware('permission:delete-page')->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -56,17 +70,28 @@ class PageController extends Controller
                     $editUrl = route('page.edit', $page->id);
                     $deleteUrl = route('page.destroy', $page->id);
 
-                    return '
-                        <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
-                            <i class="fa fa-edit"></i>
-                        </a>
-                        <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $page->id . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </form>
-                    ';
+                    $buttons = '';
+
+                    if (auth()->user()->can('edit-page')) {
+                        $buttons .= '
+                            <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
+                                <i class="fa fa-edit"></i>
+                            </a>
+                        ';
+                    }
+
+                    if (auth()->user()->can('delete-page')) {
+                        $buttons .= '
+                            <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $page->id . '" style="display:inline-block;">
+                                ' . csrf_field() . method_field('DELETE') . '
+                                <button type="submit" class="btn btn-danger btn-sm deleteItem">
+                                    <i class="fa fa-trash"></i>
+                                </button>
+                            </form>
+                        ';
+                    }
+
+                    return $buttons ?: '<span class="text-muted">No Actions</span>';
                 })
 
                 ->rawColumns(['action'])
