+++ app/Http/Controllers/Admin/VideoController.php
@@ -9,6 +9,20 @@ use Yajra\DataTables\Facades\DataTables;
 
 class VideoController extends Controller
 {
+    public $user;
+    
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = auth()->guard('web')->user();
+            return $next($request);
+        });
+
+        $this->middleware('permission:view-video')->only('index');
+        $this->middleware('permission:create-video')->only(['create', 'store']);
+        $this->middleware('permission:edit-video')->only(['edit', 'update']);
+        $this->middleware('permission:delete-video')->only('destroy');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -22,21 +36,31 @@ class VideoController extends Controller
                 ->addColumn('created_at', function ($banner) {
                     return $banner->created_at->format('d M Y'); 
                 })
-                ->addColumn('action', function(Video $video) {
+                ->addColumn('action', function (Video $video) {
                     $editUrl = route('video.edit', $video->id);
                     $deleteUrl = route('video.destroy', $video->id);
 
-                    return '
-                        <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
-                            <i class="fa fa-edit"></i>
-                        </a>
-                        <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $video->id . '" style="display:inline-block;">
-                            ' . csrf_field() . method_field('DELETE') . '
-                            <button type="submit" class="btn btn-danger btn-sm deleteItem">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </form>
-                    ';
+                    $buttons = '';
+
+                    if (auth()->user()->can('edit-video')) {
+                        $buttons .= '
+                            <a href="' . $editUrl . '" class="btn btn-sm btn-primary text-white">
+                                <i class="fa fa-edit"></i>
+                            </a>
+                        ';
+                    }
+
+                    if (auth()->user()->can('delete-video')) {
+                        $buttons .= '
+                            <form action="' . $deleteUrl . '" method="POST" class="delete-form" data-id="' . $video->id . '" style="display:inline-block;">
+                                ' . csrf_field() . method_field('DELETE') . '
+                                <button type="submit" class="btn btn-danger btn-sm deleteItem">
+                                    <i class="fa fa-trash"></i>
+                                </button>
+                            </form>
+                        ';
+                    }
+                    return $buttons ?: '<span class="text-muted">No Actions</span>';
                 })
                 ->rawColumns(['status', 'action'])
                 ->make(true);
